+++ hklpy2/blocks/sample.py
@@ -63,8 +63,8 @@ class Sample:
         self.core = core
         self.lattice = lattice
         self.U = IDENTITY_MATRIX_3X3
+        # Consider: UB = self.U @ self.lattice.B
         self.UB = ((2 * math.pi / self.lattice.a) * np.array(self.U)).tolist()
-        # TODO: reciprocal_lattice
         self.reflections = ReflectionsDict()
 
     def __repr__(self):